+++ RELEASES.md
@@ -3,4 +3,4 @@
 ## 1.4.0, 11 July 2016
 
 - shortcut for configuring stage variables during deployment
-
+- you can now provide a handler for unsupported event formats, and invoke the same Lambda from a source that's not API Gateway